Moved add building dropdown into the tile details component.
